Keep schedule datastructures with schedule
